Use response formatting to force the LLM to properly format memory generation output; Replaced sashabaranov/go-openai with openai/openai-go/v2, the official lib, as it does less magic around with the response format (as well as being official).

+++ backend/model/preferences/preferences.go
@@ -14,7 +14,7 @@ type Preferences struct {
 	// Memories
 	MemoriesModelId       *int    `json:"memoriesModelId"`
 	MemoriesInstructionId *int    `json:"memoriesInstructionId"`
-	MemoryMinP            float32 `json:"memoryMinP"`
+	MemoryMinP            float64 `json:"memoryMinP"`
 	MemoryTriggerAfter    int     `json:"memoryTriggerAfter"`
 	MemoryWindowSize      int     `json:"memoryWindowSize"`
 }
